move money SA types to deprecated namespace
